pyproject.toml
[project]
name = "aiap-22-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.1.0",
    "missingpy>=0.2.0",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pip>=25.3",
    "pyampute>=0.0.3",
    "pytest>=9.0.2",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.44",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
